Revert `int` -> `long`

+++ src/main/java/gregtech/api/util/OverclockCalculator.java
@@ -399,8 +399,8 @@ public class OverclockCalculator {
         overclocks = Math.max(overclocks, 0);
 
         // Split overclocks into heat-based and regular overclocks.
-        long heatOverclocks = Math.min(heatOC ? (machineHeat - recipeHeat) / HEAT_OVERCLOCK_THRESHOLD : 0, overclocks);
-        long regularOverclocks = overclocks - heatOverclocks;
+        int heatOverclocks = Math.min(heatOC ? (machineHeat - recipeHeat) / HEAT_OVERCLOCK_THRESHOLD : 0, overclocks);
+        int regularOverclocks = overclocks - heatOverclocks;
 
         // Adjust power consumption and processing time based on overclocks.
         calculatedConsumption = (long) Math.ceil(recipePower * GTUtility.powInt(eutIncreasePerOC, overclocks));
